requirements/base.txt
# Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# WebSocket & Real-time Support
channels==4.0.0
channels-redis==4.1.0

# Notification Dependencies
requests==2.31.0
python-telegram-bot==20.7

# Database
psycopg2-binary==2.9.9  # For production PostgreSQL
dj-database-url==2.1.0

# Caching & Session
django-redis==5.4.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3

# Health Checks & Monitoring
django-health-check==3.17.0
psutil==5.9.6

# Background Tasks
django-celery-beat==2.5.0

# Development Tools (in development.txt)
# pytest-django==4.7.0
# pytest-cov==4.1.0
# factory-boy==3.3.0
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0
